requirements.txt
kivy
bs4
requests
pillow